Move CEPiOS exclusive fixture to its scope
