+++ CEPSearch/CEPSearchTests/Shared Helpers/Fixtures.swift
@@ -6,7 +6,6 @@
 //
 
 import CEPSearch
-import CEPiOS
 
 func makeCEPDetailsFixture(cep: String = "00000-000",
                            street: String = "Default Street",
@@ -22,15 +21,4 @@ func makeCEPDetailsFixture(cep: String = "00000-000",
                       city: city,
                       state: state
     )
-}
-
-func makeCEPDetailsViewDataFixture(cepText: String = "CEP: 12345-678",
-                                   addressText: String = "Address: Example Street, Apt 101",
-                                   districtText: String = "District: Example District",
-                                   cityStateText: String = "City/State: Example City, EX"
-) -> CEPDetailsViewData {
-    return CEPDetailsViewData(cepText: cepText,
-                              addressText: addressText,
-                              districtText: districtText,
-                              cityStateText: cityStateText)
 }
